// youtube-client-app/src/app/youtube/models/video-item.model.ts
export interface VideoItem {
  kind?: string;
  etag?: string;
  id: string;
  snippet?: Snippet;
  statistics?: Statistics;
  title?: string;
  description?: string;
  imageUrl?: string;
  videoUrl?: string;
  creationDate?: Date;
  isCustom?: boolean;
}

export interface Snippet {
  publishedAt: string;
  channelId?: string;
  title: string;
  description: string;
  thumbnails: Thumbnails;
  channelTitle?: string;
  tags?: string[];
  categoryId?: string;
  liveBroadcastContent?: string;
  localized?: Localized;
  defaultAudioLanguage?: string;
}

interface Thumbnails {
  default?: Thumbnail;
  medium: Thumbnail;
  high?: Thumbnail;
  standard: Thumbnail;
  maxres?: Thumbnail;
}

interface Thumbnail {
  url: string;
  width?: number;
  height?: number;
}

interface Localized {
  title: string;
  description: string;
}

interface Statistics {
  viewCount: string;
  likeCount: string;
  dislikeCount: string;
  favoriteCount: string;
  commentCount: string;
}